Modificado ex05 com mais exemplos e comentário

+++ modulo0/ex03/average_words.h
@@ -3,6 +3,6 @@
 
 // Média de 2 Inteiros previamente convertidos de String
 
-int average_words(char *str1, char *str2) {
-    return (string_to_int(str1) + string_to_int(str2))/2;
+float average_words(char *str1, char *str2) {
+    return (float)(string_to_int(str1) + string_to_int(str2))/2;
 }
